src/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Django==5.1
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
psycopg2-binary==2.9.9
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
